# flutter_application_1/1XL-Infra-Office-Asset-Tracking-Maintenance-System-main (1)/1XL-Infra-Office-Asset-Tracking-Maintenance-System-main/server/openapi.yaml
openapi: 3.1.0
info:
  title: 1XL Asset Tracker — Mobile API
  version: 1.0.0
  description: |
    REST API consumed by the Flutter companion app for the 1XL Asset Tracker
    platform. Implements every P0 endpoint from `questionnaire.md` Section 5,
    plus photo upload, recovery, audit verification, and reference lookups.

    ### Auth model
    All requests use the standard `Authorization: Bearer <token>` scheme.
    The bearer token is one of:
    - **API key** — `<orgslug>_<live|test>_<24chars>`, e.g. `1xl_live_K8j3pQrX9bR2sN7wHzVc1qLm`.
      Issued by an org admin in the web UI.
    - **Access JWT** — short-lived per-user token from `POST /auth/login`.
    The server distinguishes them by shape. `/auth/login`, `/auth/refresh`,
    `/auth/logout` only accept an API key. `/auth/me` only accepts a user JWT.
    Everything else accepts either.

    ### Versioning
    Additive changes within `/api/v1/*`. Breaking changes get `/api/v2/*`.

    ### Idempotency
    All `POST` endpoints accept `Idempotency-Key`. Same key + same body = cached
    response for 24h. Same key + different body = `409 CONFLICT`.
servers:
  - url: https://asset.thebizzfly.com
    description: Production
  - url: http://localhost:4000
    description: Local development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        `Authorization: Bearer <token>` where `<token>` is either:
          - an API key (`<orgslug>_<live|test>_<24chars>`) — for /auth/login/refresh/logout and server-to-server calls, or
          - a user JWT (HS256) — for user-scoped calls (e.g. /auth/me).
  parameters:
    RequestId:
      name: X-Request-Id
      in: header
      schema: { type: string, maxLength: 128 }
      description: Optional client-supplied id mirrored back in the response and any error envelope.
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      schema: { type: string, minLength: 16, maxLength: 128 }
      description: Required for safe POST retries. 24h TTL.
    AssetId:
      name: assetId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    PhotoId:
      name: photoId
      in: path
      required: true
      schema: { type: string, format: uuid }
  schemas:
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code:
              type: string
              enum:
                - UNAUTHENTICATED
                - TOKEN_EXPIRED
                - INVALID_API_KEY
                - FORBIDDEN
                - INSUFFICIENT_SCOPE
                - NOT_FOUND
                - ASSET_NOT_FOUND
                - REPAIR_NOT_FOUND
                - PHOTO_NOT_FOUND
                - CYCLE_NOT_FOUND
                - INVALID_QR_PAYLOAD
                - VALIDATION_FAILED
                - CONFLICT
                - RATE_LIMITED
                - INTERNAL_ERROR
            message:    { type: string }
            requestId:  { type: string }
            fieldErrors:
              type: array
              items:
                type: object
                required: [field, message]
                properties:
                  field:   { type: string }
                  message: { type: string }
            details: { }
...
    Asset:
      type: object
      properties:
        id:             { type: string, format: uuid }
        assetTag:       { type: string }
        name:           { type: string }
        type:           { type: string, enum: [furniture, it_equipment, vehicle, electronics, office_equipment, hvac, infrastructure, other] }
        category:       { type: string }
        brand:          { type: string }
        model:          { type: string }
        serialNumber:   { type: string }
        status:
          type: string
          enum: [available, allocated, in_use, under_maintenance, retired, disposed, dead]
          description: |
            **CANONICAL LIST (Section 4.3 reconciliation).** Mobile must adopt these
            seven values exactly. The legacy 5-status mobile list (`assigned`,
            `under_repair`, `lost`) maps as: `assigned`→`allocated`,
            `under_repair`→`under_maintenance`, `lost`→`dead`.
        description:    { type: string }
        purchaseDate:   { type: string, format: date }
        purchaseCost:   { type: number }
        currency:       { type: string, nullable: true }
        warrantyStart:  { type: string, format: date, nullable: true }
        warrantyEnd:    { type: string, format: date, nullable: true }
        imageUrl:       { type: string, nullable: true }
        imageUrls:
          type: array
          items: { type: string }
        organizationId: { type: string, format: uuid }
        createdAt:      { type: string, format: date-time }
        updatedAt:      { type: string, format: date-time }
    AssetEnvelope:
      type: object
      required: [asset, permittedActions]
      properties:
        asset:           { $ref: '#/components/schemas/Asset' }
        location:        { type: object, nullable: true }
        department:      { type: object, nullable: true }
        vendor:          { type: object, nullable: true }
        currentAllocation: { type: object, nullable: true }
        currentAssignee: { type: object, nullable: true }
        auditContext:
          type: object
          nullable: true
          properties:
            cycleId:                { type: string, format: uuid }
            cycleName:              { type: string }
            cycleStartsAt:          { type: string, format: date-time }
            cycleEndsAt:            { type: string, format: date-time, nullable: true }
            expectedLocationId:     { type: string, nullable: true }
            expectedAssigneeId:     { type: string, nullable: true }
            currentVerification:
              type: object
              nullable: true
              properties:
                id:        { type: string, format: uuid }
                result:    { type: string, enum: [verified, flagged] }
                createdAt: { type: string, format: date-time }
        permittedActions:
          type: array
          items:
            type: string
            enum: [add_photo, log_maintenance, raise_repair, update_repair, mark_recovery, verify_audit]
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Validation:
      description: Request validation failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
paths:
...
  /api/v1/assets/lookup-by-qr:
    post:
      summary: Resolve a scanned QR payload to an asset envelope
      parameters: [{ $ref: '#/components/parameters/RequestId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [qrPayload]
              properties:
                qrPayload: { type: string, description: 'Asset id (UUID), asset tag, or scan/asset URL' }
                geoLat:    { type: number, minimum: -90,  maximum: 90 }
                geoLng:    { type: number, minimum: -180, maximum: 180 }
      responses:
        '200':
          description: Asset envelope
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AssetEnvelope' }
        '400': { $ref: '#/components/responses/Validation' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/assets/{assetId}:
    get:
      summary: Fetch one asset (envelope shape)
      parameters: [{ $ref: '#/components/parameters/AssetId' }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AssetEnvelope' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/assets/{assetId}/photos/upload-url:
    post:
      summary: Get a presigned PUT URL for a photo upload
      parameters:
        - { $ref: '#/components/parameters/AssetId' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [filename, mimeType, sizeBytes]
              properties:
                filename:  { type: string, maxLength: 200 }
                mimeType:  { type: string, pattern: '^image/(jpe?g|png|webp|heic|heif)$' }
                sizeBytes: { type: integer }
                caption:   { type: string, maxLength: 500 }
      responses:
        '200':
          description: Presigned URL + photo id
          content:
            application/json:
              schema:
                type: object
                properties:
                  photoId:           { type: string, format: uuid }
                  uploadUrl:         { type: string }
                  uploadMethod:      { type: string, const: PUT }
                  uploadHeaders:     { type: object, additionalProperties: { type: string } }
                  storagePath:       { type: string }
                  bucket:            { type: string }
                  expiresInSeconds:  { type: integer }
                  maxSizeBytes:      { type: integer }
  /api/v1/assets/{assetId}/photos/{photoId}/finalize:
    post:
      summary: Confirm a photo upload and attach it to the asset
      parameters:
        - { $ref: '#/components/parameters/AssetId' }
        - { $ref: '#/components/parameters/PhotoId' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                caption: { type: string, maxLength: 500 }
      responses:
        '200':
          description: Photo finalized and attached
          content:
            application/json:
              schema:
                type: object
                properties:
                  photo:
                    type: object
                    properties:
                      id:          { type: string, format: uuid }
                      url:         { type: string }
                      caption:     { type: string, nullable: true }
                      mimeType:    { type: string }
                      sizeBytes:   { type: integer }
                      createdAt:   { type: string, format: date-time }
                      finalizedAt: { type: string, format: date-time }
        '404': { $ref: '#/components/responses/NotFound' }

  # ─── Maintenance ──────────────────────────────────────────────────────────
  /api/v1/assets/{assetId}/maintenance:
    get:
      summary: List maintenance entries for an asset
      parameters: [{ $ref: '#/components/parameters/AssetId' }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { type: object } }
                  total: { type: integer }
    post:
      summary: Log a maintenance entry
      parameters:
        - { $ref: '#/components/parameters/AssetId' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [taskTypeId, type]
              properties:
                taskTypeId:    { type: string }
                type:          { type: string, enum: [preventive, corrective] }
                scheduledDate: { type: string, format: date-time }
                completedDate: { type: string, format: date-time }
                technicianId:  { type: string, format: uuid, nullable: true }
                cost:          { type: number, minimum: 0 }
                notes:         { type: string }
                checklist:     { type: array, items: { type: string } }
                status:        { type: string, enum: [scheduled, in_progress, completed, overdue, cancelled] }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  maintenance: { type: object }
...